% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_codon.R
\name{change.codon}
\alias{change.codon}
\title{Generates a silent mutation to a codon.}
\usage{
change.codon(cps, codpos, sq)
}
\value{
A DNA sequence expressing an ORF with an alternative set of equivalent codons.
}
\description{
It creates a silent mutation to a codon at the specified position.
}
\author{
Máximo Sánchez-Aragón
}
