% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_alignment.R
\name{replace.alignment}
\alias{replace.alignment}
\title{Introduces silent mutation in the alignment regions of two sequences}
\usage{
replace.alignment(
  cdna.file,
  rnai.file,
  gcode = NULL,
  codon.length = 3,
  gap.opening = 100
)
}
\arguments{
\item{cdna.file}{FASTA file containing the cDNA sequence (larger sequence).}

\item{rnai.file}{FASTA file containing the RNAi sequence (shorter sequence).}

\item{gcode}{Genetic code used. If \code{NULL}, the \code{GENETIC_CODE} object from the BiocManager package is used.}

\item{codon.length}{The length of a codon, typically 3}

\item{gap.opening}{Indel penalty for ClustalW.}
}
\value{
Returns a genetic sequence with silent mutations at the alignment region.
}
\description{
It introduces silent mutations exclusively in the alignment region of two sequences, typically a cDNA sequence and a RNAi sequence.
}
\examples{
rnai.file <- "inst/extdata/panos_rnai"
cdna.file <- "inst/extdata/panos_cdna"
replace.alignment(cdna.file,rnai.file)
}
\author{
Máximo Sánchez-Aragón.
}
